Switched to using long for distance vector

+++ include/bellman_ford.h
@@ -6,7 +6,7 @@
 bool bellman_ford(Graph&,
 				  Vertex s,
 				  CostPropertyMap &costs,
-				  std::vector<int> &dist,
+				  std::vector<long> &dist,
 				  std::vector<Vertex> &pred);
